re-organise tests to help contributors get started.
