+++ lib/index.js
@@ -1,16 +1,16 @@
 module.exports = {
   CONNECT: require('./connection'),
-  CREATE: require('./create'),
-  READ:   require('./read'),
-  UPDATE: require('./update'),
-  DELETE: require('./delete'),
-  STATS: require('./stats'),
-  SEARCH: require('./search'),
+  CREATE:  require('./create'),
+  READ:    require('./read'),
+  UPDATE:  require('./update'),
+  DELETE:  require('./delete'),
+  STATS:   require('./stats'),
+  SEARCH:  require('./search'),
   connect: require('./connection'),
-  create: require('./create'),
-  read:   require('./read'),
-  update: require('./update'),
-  delete: require('./delete'),
-  stats: require('./stats'),
-  search: require('./search')
+  create:  require('./create'),
+  read:    require('./read'),
+  update:  require('./update'),
+  delete:  require('./delete'),
+  stats:   require('./stats'),
+  search:  require('./search')
 }
