+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "esta",
-  "version": "4.0.1",
+  "version": "4.0.2",
   "description": "The Simplest ElasticSearch Node.js Client",
   "main": "lib/index.js",
   "directories": {
